Allow to build dlls
The BUILD_SHARED_LIBS option can now turn on shared libs also for win32 target.

Every non-template class or function must export it's symbols.
A macro MLPACK_API is added to prereqs.hpp for this.

MLPACK_STATIC must be defined when using the static lib.

+++ src/mlpack/methods/det/dt_utils.cpp
@@ -192,7 +192,7 @@ DTree* mlpack::det::Trainer(arma::mat& dataset,
   {
     // Break up data into train and test sets.
     size_t start = fold * testSize;
-    size_t end = std::min((fold + 1) * testSize, (size_t) cvData.n_cols);
+    size_t end = std::min((size_t) (fold + 1) * testSize, (size_t) cvData.n_cols);
 
     arma::mat test = cvData.cols(start, end - 1);
     arma::mat train(cvData.n_rows, cvData.n_cols - test.n_cols);
